lots of changes
modified util, home_page and deadlines. made the code more lightweight, made the second modal for deadlinespage which supposedly is able to add to the transaction list in homepage. dunno if it is able to remove the deadline items from deadlines yet, but it should be able to udate the balance.
TODO:
- I might have to correct it to not update the balance if the payment method is not the uni bank account.
- need to use the same modal in details in a way that only makes me pay the single transaction, as in, UI should see it as if I had selected only one transaction and decided to pay that.

After that I think the STUDENT UI is alright, and we can start with PARENT UI.