committed for elena to run 2day
I don't remember my last commit but this one should run, the code for homepage is shortened by a lot bc I have moved some stuff to util! I suggest we use this as it is more efficient and works with my other page, and we put the latest additions in the other branches to my version of home_page/util if necessary.